requirements.txt
olefile==0.44
Pillow==4.1.1
wheel==0.24.0
